chore: refactor consumer hook
